Add support for python objects
If a value can't be converted from python to Lisp, it is
put into a dict and a handle is returned. This is put into
a `python-object` struct. When converting back to python,
this struct is turned into a reference to the entry in the dict.

At the moment objects put into this dict store are never released.